requirements.txt
streamlit
python-docx
python-pptx
pandas
PyMuPDF
markdown
openai
faiss-cpu
sentence-transformers
tiktoken
python-dotenv
google-generativeai
nltk
chromadb
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langgraph